elf: Move Elf64File to separate module
In the effort of splitting up the ELF parser code, move the Elf64File
struct to its own module.
